pyproject.toml
[tool.poetry]
name = "my-railbot-app"
version = "0.1.0"
description = ""
authors = ["Surajit Ghorai <ghoraisurajit2021@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^0.1.13"
langchain-openai = "^0.1.1"
python-dotenv = "^1.0.1"
sentence-transformers = "^2.6.1"
chromadb = "^0.4.24"
tavily-python = "^0.3.1"
langchainhub = "^0.1.15"
langsmith = "^0.1.37"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
duckduckgo-search = "^5.2.1"
unstructured = "^0.12.6"
pypdf = "^4.1.0"
langchain-google-genai = "^1.0.1"
docx2txt = "^0.8"
python-docx = "^1.1.0"
python-pptx = "^0.6.23"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
markdown = "^3.6"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
